extract response_objects format methods into their own lib
