--- app/spec/support/fixtures/argyle/request_user.json
@@ -1,15 +1,0 @@
-{
-    "id": "0195441c-5a5f-7d86-3be1-fa5797a441a6",
-    "created_at": "2025-02-26T21:14:23.712349Z",
-    "items_connected": [
-        "item_000024123",
-        "item_000020392",
-        "item_000041078"
-    ],
-    "employers_connected": [
-        "Uber Driver",
-        "Lyft Driver"
-    ],
-    "external_metadata": {},
-    "external_id": "testing"
-}
